Adding Router.RunAction, the ability to run an action on the dispatch loop without publishing an event Simplified the IRouter interface, terminal event handler now goes via the ctor, not IRouter.AddTerminalErrorHandler() Remove IRouter.CreateModelRouter(), to do this now you just new up a Router<TModel>() and choose a ctor for your needs The various router type can now be found in Router.cs, Router`1.cs and Router`2.cs Bumpted version to 0.5.0-alpha
