+++ src/Esp.Net.Tests/RouterTests.ModelSubRouter.cs
@@ -15,7 +15,7 @@ namespace Esp.Net
             public override void SetUp()
             {
                 base.SetUp();
-                _modelRouter = _router.CreateModelRouter<TestModel, SubTestModel>(_model1.Id, m => m.SubTestModel);
+                _modelRouter = new Router<TestModel, SubTestModel>(_model1.Id, _router, m => m.SubTestModel);
             }
 
             [Test]
@@ -35,14 +35,6 @@ namespace Esp.Net
                 _modelRouter.GetModelObservable().Observe(m => receivedModelCount++);
                 _modelRouter.PublishEvent(new Event1());
                 receivedModelCount.ShouldBe(1);
-            }
-
-            [Test]
-            public void CanBroadcastProxiedEvent()
-            {
-                _modelRouter.BroadcastEvent(new Event1());
-                _model1EventProcessor.Event1Details.NormalStage.ReceivedEvents.Count.ShouldBe(1);
-                _model2EventProcessor.Event1Details.NormalStage.ReceivedEvents.Count.ShouldBe(1);
             }
         }
     }
