+++ src/Esp.Net.Tests/RouterTests.RegisterModel.cs
@@ -10,31 +10,31 @@ namespace Esp.Net
             [Test]
             public void ThrowsIfModelIdNull()
             {
-                Assert.Throws<ArgumentNullException>(() => _router.RegisterModel(null, new object()));
+                Assert.Throws<ArgumentNullException>(() => _router.AddModel(null, new object()));
             }
 
             [Test]
             public void ThrowsIfPreEventProcessorNull()
             {
-                Assert.Throws<ArgumentNullException>(() => _router.RegisterModel(Guid.NewGuid(), new object(), (IPreEventProcessor<object>)null));
+                Assert.Throws<ArgumentNullException>(() => _router.AddModel(Guid.NewGuid(), new object(), (IPreEventProcessor<object>)null));
             }
 
             [Test]
             public void ThrowsIfPostEventProcessorNull()
             {
-                Assert.Throws<ArgumentNullException>(() => _router.RegisterModel(Guid.NewGuid(), new object(), (IPostEventProcessor<object>)null));
+                Assert.Throws<ArgumentNullException>(() => _router.AddModel(Guid.NewGuid(), new object(), (IPostEventProcessor<object>)null));
             }
 
             [Test]
             public void ThrowsIfModelNull()
             {
-                Assert.Throws<ArgumentNullException>(() => _router.RegisterModel(Guid.NewGuid(), (object)null));
+                Assert.Throws<ArgumentNullException>(() => _router.AddModel(Guid.NewGuid(), (object)null));
             }
 
             [Test]
             public void ThrowsIfModelAlreadyRegistered()
             {
-                Assert.Throws<ArgumentException>(() => _router.RegisterModel(_model1.Id, new TestModel()));
+                Assert.Throws<ArgumentException>(() => _router.AddModel(_model1.Id, new TestModel()));
             }
         }
     }
