+++ src/Esp.Net.Tests/packages.config
@@ -1,12 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-Sample license text.
--->
 <packages>
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
   <package id="Rx-Core" version="2.2.5" targetFramework="net451" />
   <package id="Rx-Interfaces" version="2.2.5" targetFramework="net451" />
   <package id="Rx-Linq" version="2.2.5" targetFramework="net451" />
+  <package id="Rx-Main" version="2.2.5" targetFramework="net451" />
+  <package id="Rx-PlatformServices" version="2.2.5" targetFramework="net451" />
+  <package id="Rx-Testing" version="2.2.5" targetFramework="net451" />
   <package id="Shouldly" version="2.5.0" targetFramework="net45" />
 </packages>
