+++ src/Esp.Net/Esp.Net.csproj
@@ -44,6 +44,8 @@
     <Compile Include="..\AssemblyInfo.Common.cs">
       <Link>Properties\AssemblyInfo.Common.cs</Link>
     </Compile>
+    <Compile Include="IRouter.`1.cs" />
+    <Compile Include="ITerminalErrorHandler.cs" />
     <Compile Include="RouterExt.ObserveEventsOn.cs" />
     <Compile Include="CurrentThreadDispatcher.cs" />
     <Compile Include="Disposables\CollectionDisposable.cs" />
@@ -65,8 +67,7 @@
     <Compile Include="Meta\ModelEventObservations.cs" />
     <Compile Include="Meta\ModelsEventsObservations.cs" />
     <Compile Include="ModelChangedEvent.cs" />
-    <Compile Include="ModelRouter.cs" />
-    <Compile Include="Router.ModelSpeific.cs" />
+    <Compile Include="Router.`1.cs" />
     <Compile Include="ObservationStage.cs" />
     <Compile Include="ObserveBaseEventAttribute.cs" />
     <Compile Include="ObserveEventAttribute.cs" />
@@ -84,10 +85,10 @@
     <Compile Include="Reactive\ModelObserver.cs" />
     <Compile Include="Reactive\ModelSubject.cs" />
     <Compile Include="Router.cs" />
-    <Compile Include="Router.ModelEntry.cs" />
+    <Compile Include="Router.ModelRouter.cs" />
     <Compile Include="Router.State.cs" />
     <Compile Include="Router.Status.cs" />
-    <Compile Include="SubModelRouter.cs" />
+    <Compile Include="Router.`2.cs" />
     <Compile Include="Utils\Guard.cs" />
     <Compile Include="Utils\ReflectionExt.cs" />
     <Compile Include="Utils\ReflectionHelper.cs" />
