+++ src/Esp.Net/ITerminalErrorHandler.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Esp.Net
+{
+    public interface ITerminalErrorHandler
+    {
+        void OnError(Exception exception); 
+    }
+}
